refactor(moonpool-sim): move LocalityInfo and DomainLevel to a top-level module
The simulation engine needs the locality vocabulary to shape network faults,
but the sim module must not import from the runner module. Move the two
vocabulary types to a new crate-level moonpool-sim/src/locality.rs that both
layers can depend on; LocalityConfig (layout) and MachineRegistry (registry)
stay in runner::locality since they are runner-side concerns.

Crate-root re-exports are unchanged, so external call sites keep working.
