+++ moonpool-sim/src/runner/locality.rs
@@ -1,10 +1,15 @@
-//! Failure-domain locality for correlated fault injection.
+//! Failure-domain topology layout and the IP → locality registry.
 //!
 //! Locality models a `FoundationDB`-style Cluster → Datacenter → Zone → Machine →
 //! Process hierarchy so that collocated processes share fate. It is **orthogonal
 //! to tags** ([`super::tags`]): tags round-robin independent dimensions, while
 //! locality assigns *contiguous* hierarchical groups whose processes can be
 //! rebooted together (see [`AttritionScope`](super::process::AttritionScope)).
+//!
+//! The vocabulary types ([`DomainLevel`], [`LocalityInfo`]) live in
+//! [`crate::locality`] because the simulation engine consumes them too; this
+//! module owns the *layout* ([`LocalityConfig`]) and the *registry*
+//! ([`MachineRegistry`]), both runner-side concerns.
 //!
 //! # Example
 //!
@@ -16,74 +21,9 @@
 use std::collections::HashMap;
 use std::net::IpAddr;
 
+use crate::locality::{DomainLevel, LocalityInfo};
+
 use super::builder::ProcessCount;
-
-/// The level of a failure domain in the locality hierarchy.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum DomainLevel {
-    /// A whole datacenter.
-    Datacenter,
-    /// A zone within a datacenter.
-    Zone,
-    /// A single machine — the unit of shared fate.
-    Machine,
-}
-
-/// Resolved failure-domain locality for a single process instance.
-///
-/// Identifiers are globally unique and hierarchical (`dc1`, `dc1-z1`,
-/// `dc1-z1-m1`) so that domain queries never confuse a machine in one
-/// datacenter with a machine in another.
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct LocalityInfo {
-    datacenter: String,
-    zone: String,
-    machine: String,
-}
-
-impl LocalityInfo {
-    /// Create locality from explicit datacenter, zone, and machine ids.
-    #[must_use]
-    pub fn new(
-        datacenter: impl Into<String>,
-        zone: impl Into<String>,
-        machine: impl Into<String>,
-    ) -> Self {
-        Self {
-            datacenter: datacenter.into(),
-            zone: zone.into(),
-            machine: machine.into(),
-        }
-    }
-
-    /// The datacenter id (e.g. `dc1`).
-    #[must_use]
-    pub fn datacenter(&self) -> &str {
-        &self.datacenter
-    }
-
-    /// The zone id (e.g. `dc1-z1`).
-    #[must_use]
-    pub fn zone(&self) -> &str {
-        &self.zone
-    }
-
-    /// The machine id (e.g. `dc1-z1-m1`).
-    #[must_use]
-    pub fn machine(&self) -> &str {
-        &self.machine
-    }
-
-    /// The id at the given domain level.
-    #[must_use]
-    pub fn id_for(&self, level: DomainLevel) -> &str {
-        match level {
-            DomainLevel::Datacenter => &self.datacenter,
-            DomainLevel::Zone => &self.zone,
-            DomainLevel::Machine => &self.machine,
-        }
-    }
-}
 
 /// Configuration for laying processes out across a failure-domain topology.
 ///
@@ -287,13 +227,5 @@ mod tests {
             vec!["dc1".to_string(), "dc2".to_string()]
         );
         assert!(!reg.is_empty());
-    }
-
-    #[test]
-    fn id_for_matches_accessors() {
-        let loc = LocalityInfo::new("dc1", "dc1-z2", "dc1-z2-m3");
-        assert_eq!(loc.id_for(DomainLevel::Datacenter), loc.datacenter());
-        assert_eq!(loc.id_for(DomainLevel::Zone), loc.zone());
-        assert_eq!(loc.id_for(DomainLevel::Machine), loc.machine());
     }
 }
